feat($DelimiterConstants): 상수 .(dot) 추가
구분자에서 . 을 막기 위해 상수 . 을 추가

+++ src/main/java/calculator/View/OutputView.java
@@ -7,7 +7,7 @@ public class OutputView {
         System.out.print(OutputMessages.RESULT);
     }
 
-    public static void printResult(int result) {
+    public static void printResult(Double result) {
         System.out.println(result);
     }
 }
